+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+* Added the (sender/receiver(s)) vCards on the XML event messages
+  which are passed to Sidekiq jobs first argument
+
 ## 0.1.0
 
 * Implemented the basic MAM to Sidekiq bridge